@@ -17,8 +17,8 @@
     "scripts": {
         "build": "rimraf ./dist && tsc -p tsconfig.build.json --incremental false",
         "check-dependencies": "dependency-check package.json --missing --no-dev",
-        "check-formatting": "prettier --check \"./**/*\"",
-        "format": "prettier --write \"./**/*\"",
+        "check-formatting": "prettier --check .",
+        "format": "prettier --write .",
         "lint": "eslint ./src --ext .ts",
         "lint:ci": "eslint ./src --ext .ts --format junit -o reports/junit/eslint.xml",
         "semantic-release": "semantic-release",
@@ -29,38 +29,32 @@
         "test:e2e": "jest --config ./test/jest-e2e.config.json"
     },
     "dependencies": {
-        "rxjs": "^6.5.5"
-    },
-    "husky": {
-        "hooks": {
-            "pre-commit": "pretty-quick --staged"
-        }
+        "rxjs": "^7.3.0"
     },
     "devDependencies": {
         "@semantic-release/changelog": "^5.0.1",
         "@semantic-release/commit-analyzer": "^8.0.1",
-        "@semantic-release/github": "^7.0.5",
-        "@semantic-release/npm": "^7.0.5",
-        "@semantic-release/release-notes-generator": "^9.0.1",
-        "@types/jest": "^25.2.1",
-        "@types/node": "^12.12.35",
-        "@typescript-eslint/eslint-plugin": "^2.27.0",
-        "@typescript-eslint/parser": "^2.27.0",
+        "@semantic-release/github": "^7.2.3",
+        "@semantic-release/npm": "^7.1.3",
+        "@semantic-release/release-notes-generator": "^9.0.3",
+        "@types/jest": "^27.0.1",
+        "@types/node": "^12.20.19",
+        "@typescript-eslint/eslint-plugin": "^4.29.1",
+        "@typescript-eslint/parser": "^4.29.1",
         "condition-circle": "^2.0.2",
         "dependency-check": "^4.1.0",
-        "eslint": "^6.8.0",
-        "eslint-config-prettier": "^6.10.1",
-        "eslint-plugin-prettier": "^3.1.3",
-        "husky": "^4.2.5",
-        "jest": "^25.3.0",
-        "jest-circus": "^25.3.0",
-        "jest-html-reporter": "^3.1.0",
-        "jest-junit": "^10.0.0",
-        "prettier": "^2.0.4",
-        "pretty-quick": "^2.0.1",
+        "eslint": "^7.32.0",
+        "eslint-config-prettier": "^8.3.0",
+        "eslint-plugin-prettier": "^3.4.0",
+        "husky": "^7.0.1",
+        "jest": "^27.0.6",
+        "jest-html-reporter": "^3.4.1",
+        "jest-junit": "^12.2.0",
+        "prettier": "^2.3.2",
+        "pretty-quick": "^3.1.1",
         "rimraf": "^3.0.2",
-        "semantic-release": "^17.0.4",
-        "ts-jest": "^25.3.1",
-        "typescript": "~3.8.3"
+        "semantic-release": "^17.4.4",
+        "ts-jest": "^27.0.4",
+        "typescript": "~4.3.5"
     }
 }
